+++ it/server/src/test/java/com/walmartlabs/concord/it/server/AnsibleIT.java
@@ -445,7 +445,7 @@ public class AnsibleIT extends AbstractServerIT {
 
         // ---
         ProcessEventsApi eventsApi = new ProcessEventsApi(getApiClient());
-        List<ProcessEventEntry> events = eventsApi.list(pir.getInstanceId(), "ANSIBLE", null, null, null, null, null);
+        List<ProcessEventEntry> events = eventsApi.list(pir.getInstanceId(), "ANSIBLE", null, null, null,null, null, null);
         assertNotNull(events);
         // one pre and one post event
         assertEquals(3, events.size());